Fix query timing precision in AccountSearchHelper
Property was int and microtime(true) was cast to int, truncating
sub-second precision. The elapsed time displayed in the search
grid was off by up to 1 second. Changed to float to match
getElapsedTime(float) and GridBase which does this correctly.

+++ src/Infrastructure/Adapter/In/Web/Controllers/Helpers/Account/AccountSearchHelper.php
@@ -67,7 +67,7 @@ final class AccountSearchHelper extends HelperBase
      */
     private bool $filterOn = false;
     private bool $isAjax   = false;
-    private int  $queryTimeStart;
+    private float $queryTimeStart;
     private bool                    $isIndex;
     private ?AccountSearchFilterDto $accountSearchFilter = null;
 
@@ -86,7 +86,7 @@ final class AccountSearchHelper extends HelperBase
     ) {
         parent::__construct($application, $template, $request);
 
-        $this->queryTimeStart = (int)microtime(true);
+        $this->queryTimeStart = microtime(true);
         $this->isIndex = $this->request->analyzeString('r') ===
                          $this->acl->getRouteFor(AclActionsInterface::ACCOUNT);
         $this->setVars();
